# Fix state performance issues + add perf monitoring