--- a/packages/back/src/battle/run/entities/player/Player.ts
+++ b/packages/back/src/battle/run/entities/player/Player.ts
@@ -26,33 +26,4 @@ export const Player = (
         socket,
         teamId
     };
-
-    // const this_: Player = {
-    //     id: playerData.id,
-    //     name: playerData.name,
-    //     socket: playerData.socket,
-    //     teamId,
-    //     get characters() {
-    //         return characters;
-    //     },
-    //     toSnapshot(): PlayerSnapshot {
-    //         return {
-    //             id: this.id,
-    //             name: this.name,
-    //             charactersSnapshots: this.characters.map(c => c.toSnapshot())
-    //         };
-    //     },
-    //     updateFromSnapshot(snapshot) {
-
-    //         characters.forEach(character => character.updateFromSnapshot(
-    //             snapshot.charactersSnapshots.find(snap => snap.id === character.id)!
-    //         ));
-    //     }
-    // };
-
-    // const characters = playerData.staticCharacters.map(({ staticData, initialPosition }) =>
-    //     characterCreator(staticData, initialPosition, this_)
-    // );
-
-    // return this_;
 };