--- a/packages/back/src/battle/run/entities/team/Team.ts
+++ b/packages/back/src/battle/run/entities/team/Team.ts
@@ -16,37 +16,4 @@ export const Team = ({ id, letter }: Pick<TeamRoom, 'id' | 'letter'>) => {
         id,
         letter
     };
-
-    // const this_: Team = {
-    //     id: teamData.id,
-    //     letter: teamData.letter,
-    //     get players() {
-    //         return players;
-    //     },
-    //     get characters() {
-    //         return characters;
-    //     },
-    //     toSnapshot(): TeamSnapshot {
-    //         return {
-    //             id: this_.id,
-    //             letter: this_.letter,
-    //             playersSnapshots: players.map(p => p.toSnapshot())
-    //         };
-    //     },
-    //     updateFromSnapshot(snapshot) {
-    //         players.forEach(player => player.updateFromSnapshot(
-    //             snapshot.playersSnapshots.find(snap => snap.id === player.id)!
-    //         ));
-    //     },
-    //     clone() {
-    //         const team = Team(teamData, deps);
-    //         team.updateFromSnapshot(this_.toSnapshot());
-    //         return team;
-    //     }
-    // };
-
-    // const players = teamData.players.map(p => playerCreator(p, this_));
-    // const characters = players.flatMap(p => p.characters);
-
-    // return this_;
 };