--- a/packages/front/src/stages/battle/snapshot/battle-data.ts
+++ b/packages/front/src/stages/battle/snapshot/battle-data.ts
@@ -1,6 +1,6 @@
+import { Normalized } from '@timeflies/shared';
 import { Character } from '../entities/character/Character';
 import { Spell } from '../entities/spell/Spell';
-import { Normalized } from '../entities/normalize';
 
 export const periodList = [ 'current', 'future' ] as const;
 export type BattleDataPeriod = (typeof periodList)[ number ];