Spell action (move) effect monitored with these values:

- Before doing anything
  - unallowed tile hover action (move) => 520ms
  - allowed tile hover action (move) => 910ms
  - allowed tile click action (move) => 1250ms
- After normalizing entities & positions/areas
  - unallowed tile hover action (move) => 440ms
  - allowed tile hover action (move) => 780ms
  - allowed tile click action (move) => 1190ms

Notes:
- major part of perf issues come from view updates, and because of multiple actions dispatched on user action;
- should think of a CI tool (or github-action) which log performance regression/improvement.

fix #13